@@ -1,10 +1,14 @@
+[project]
+name = "treestamps"
+requires-python = ">=3.10"
+
 [build-system]
 requires = ["poetry-core>=1.0.0"]
 build-backend = "poetry.core.masonry.api"
 
 [tool.poetry]
 name = "treestamps"
-version = "0.4.4"
+version = "1.0.0"
 description = "Create timestamp records for recursive operations on directory trees."
 authors = ["AJ Slater <aj@slater.net>"]
 license = "GPL-3.0-only"
@@ -18,6 +22,7 @@ classifiers = [
   "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
   "Natural Language :: English",
   "Operating System :: OS Independent",
+  "Development Status :: 5 - Production/Stable",
   "Programming Language :: Python :: 3 :: Only",
   "Topic :: Software Development :: Libraries :: Python Modules",
 ]
@@ -25,7 +30,7 @@ packages = [{ include = "treestamps" }, { include = "tests", format = "sdist" }]
 include = ["NEWS.md"]
 
 [tool.poetry.dependencies]
-python = "^3.9"
+python = "^3.10"
 "ruamel.yaml" = "^0.18.5"
 termcolor = "^2.0.1"
 
@@ -110,9 +115,11 @@ exclude = [
   "typings",
   "vulture_whitelist.py"
 ]
-useLibraryCodeForTypes = true
-reportMissingImports = true
+pythonVersion = "3.10"
+pythonPlatform = "All"
 reportImportCycles = true
+reportMissingImports = true
+useLibraryCodeForTypes = true
 
 [tool.radon]
 exclude = "*~,.git/*,.mypy_cache/*,.pytest_cache/*,.ruff_cache/*,.venv/*,__pycache__/*,dist/*,node_modules/*,test-results/*,typings/*"
@@ -134,7 +141,7 @@ extend-select = ["A", "ARG", "B", "B9", "C", "C4", "C90", "D", "DJ",
 external = ["V101"]
 # format = "grouped"
 # show-source = true
-target-version = "py39"
+target-version = "py310"
 task-tags = ["TODO", "FIXME", "XXX", "http", "HACK"]
 
 [tool.ruff.per-file-ignores]
